Initial round of refactoring.
