+++ src/main/java/org/springframework/data/neo4j/repository/query/CypherQueryCreator.java
@@ -256,7 +256,7 @@ final class CypherQueryCreator extends AbstractQueryCreator<QueryFragmentsAndPar
 	@Override
 	protected QueryFragmentsAndParameters complete(@Nullable Condition condition, Sort sort) {
 
-		QueryFragmentsAndParameters.QueryFragments queryFragments = createQueryFragments(condition, sort);
+		QueryFragments queryFragments = createQueryFragments(condition, sort);
 
 		Map<String, Object> convertedParameters = this.boundedParameters.stream()
 				.collect(Collectors.toMap(p -> p.nameOrIndex, p -> parameterConversion.apply(p.value, p.conversionOverride)));
@@ -264,8 +264,8 @@ final class CypherQueryCreator extends AbstractQueryCreator<QueryFragmentsAndPar
 	}
 
 	@NonNull
-	private QueryFragmentsAndParameters.QueryFragments createQueryFragments(@Nullable Condition condition, Sort sort) {
-		QueryFragmentsAndParameters.QueryFragments queryFragments = new QueryFragmentsAndParameters.QueryFragments();
+	private QueryFragments createQueryFragments(@Nullable Condition condition, Sort sort) {
+		QueryFragments queryFragments = new QueryFragments();
 
 		// all the ways we could query for
 		Node startNode = Cypher.node(nodeDescription.getPrimaryLabel(), nodeDescription.getAdditionalLabels())
